% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captioner.R
\name{chris_note}
\alias{chris_note}
\title{Add a default note to the plot caption}
\usage{
chris_note(
  text = getOption("chris.caption", "caprosser.com"),
  sep = " | ",
  append = TRUE
)
}
\arguments{
\item{text}{Caption text}

\item{sep}{Separator}

\item{append}{Append to exising caption or not}
}
\description{
Add a default note to the plot caption
}
