% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{chris_palette}
\alias{chris_palette}
\title{Discrete Chris palette (12 colours)
Slightly tweaked version of https://tsitsul.in/blog/coloropt/}
\format{
An object of class \code{character} of length 12.
}
\usage{
chris_palette
}
\description{
Discrete Chris palette (12 colours)
Slightly tweaked version of https://tsitsul.in/blog/coloropt/
}
\keyword{datasets}
