% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{coloropt_palettes}
\alias{coloropt_palettes}
\alias{opt_norm_palette_6}
\alias{opt_norm_palette_12}
\alias{opt_bright_palette}
\alias{opt_dark_palette}
\alias{opt_fancy_palette}
\alias{opt_tarnish_palette}
\alias{chris_palette}
\title{Optimal qualitative colour palettes}
\format{
An object of class \code{character} of length 7.

An object of class \code{character} of length 13.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.

An object of class \code{character} of length 13.
}
\usage{
opt_norm_palette_6

opt_norm_palette_12

opt_bright_palette

opt_dark_palette

opt_fancy_palette

opt_tarnish_palette

chris_palette
}
\description{
Palettes taken from \url{https://tsitsul.in/blog/coloropt/}.
\code{chris_palette} is just a reordered version of \code{opt_norm_palette_12} and is used as the default palette for \code{theme_chris}
}
\keyword{datasets}
