% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_source_sans.R
\name{register_source_sans}
\alias{register_source_sans}
\title{Register Source Sans 3 for plotting (strict)}
\usage{
register_source_sans()
}
\description{
Called automatically on package load; errors if fonts are missing.
You can also call it manually to re-register after clearing devices.
}
