% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic.R
\name{theme_academic}
\alias{theme_academic}
\title{A plain version of theme_chris}
\usage{
theme_academic(
  base_family = "Source Sans 3",
  base_size = 11,
  ...,
  bg_panel = "white",
  bg_plot = "white",
  text_col = "black",
  grid_col = "grey70",
  strip_bg = NA,
  strip_col = NA,
  box_col = "grey30"
)
}
\arguments{
\item{base_family}{base font family}

\item{base_size}{text size}

\item{...}{Forwarded to \code{\link[=theme_chris]{theme_chris()}} (e.g., \code{title_family}, \code{base_size}, etc.)}

\item{bg_panel}{panel background colour}

\item{bg_plot}{plot background colour}

\item{text_col}{text colour}

\item{grid_col}{gridlines colour}

\item{strip_bg}{facet strip background colour}

\item{strip_col}{facet strip outline colour}

\item{box_col}{plot border colour}
}
\description{
Just calls \code{\link[=theme_chris]{theme_chris()}} with plain defaults (white bg, black text, grey grid).
}
