% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_chris.R
\name{theme_chris}
\alias{theme_chris}
\title{Custom ggplot theme}
\usage{
theme_chris(
  base_family = "Source Sans 3",
  base_size = 11,
  bg_panel = "#f8fafc",
  bg_plot = "#f8fafc",
  text_col = "#0F172A",
  grid_col = "#d4e0ec",
  strip_bg = "#d4e0ec",
  strip_col = "#d4e0ec",
  box_col = "#d4e0ec"
)
}
\arguments{
\item{base_family}{base font family}

\item{base_size}{text size}

\item{bg_panel}{panel background colour}

\item{bg_plot}{plot background colour}

\item{text_col}{text colour}

\item{grid_col}{gridlines colour}

\item{strip_bg}{facet strip background colour}

\item{strip_col}{facet strip outline colour}

\item{box_col}{plot border colour}
}
\description{
My pretty ggplot theme, based on \code{theme_bw()}, with theme-level palettes (ggplot2 >= 4.0.0)
}
