% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_chris.R
\name{use_chris}
\alias{use_chris}
\title{Set theme + defaults for the current session}
\usage{
use_chris(caption = "caprosser.com", set_geom_defaults = TRUE)
}
\arguments{
\item{caption}{default caption text (or NULL to skip)}

\item{set_geom_defaults}{Whether to use use_chris_geom_defaults or not}
}
\description{
Set theme + defaults for the current session
}
