% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_geom_defaults.R
\name{use_chris_geom_defaults}
\alias{use_chris_geom_defaults}
\title{Update geom defaults to nicer defaults
Use \code{reset_geom_defaults()} to reset}
\usage{
use_chris_geom_defaults()
}
\description{
Update geom defaults to nicer defaults
Use \code{reset_geom_defaults()} to reset
}
